Final cleanups on code, comments, indentation..

+++ db/schema.sql
@@ -3,24 +3,24 @@ CREATE DATABASE employee_DB;
 USE employee_DB;
 
 CREATE TABLE departments (
-    id INTEGER NOT NULL AUTO_INCREMENT,
+    id INT NOT NULL AUTO_INCREMENT,
     name VARCHAR(30) NOT NULL,
     PRIMARY KEY(id)
 ); 
 
 CREATE TABLE roles (
-    id INTEGER NOT NULL AUTO_INCREMENT,
+    id INT NOT NULL AUTO_INCREMENT,
     title VARCHAR(30) NOT NULL, 
     salary VARCHAR (30) NOT NULL,
-    department_id INTEGER(9) NOT NULL,
+    department_id INT(9) NOT NULL,
     PRIMARY KEY (id)
 ); 
 
 CREATE TABLE employees (
-    id INTEGER NOT NULL AUTO_INCREMENT,
+    id INT NOT NULL AUTO_INCREMENT,
     first_name VARCHAR(30) NOT NULL,
     last_name VARCHAR(30) NOT NULL,
-    role_id INTEGER(9) NOT NULL,
-    manager_id INTEGER(9) NOT NULL,
+    role_id INT(9) NOT NULL,
+    manager_id INT(9) NOT NULL,
     PRIMARY KEY(id)
 );
